harmony_service/pip_requirements.txt
# Open source packages available from PyPI:
harmony-service-lib ~= 1.0.12
pystac ~= 0.5.3
